requirements.txt
streamlit
python-dotenv
pypdf
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-text-splitters
langchain-groq
voyageai
